Formatting

+++ benchmarks/benchmark_lennard_jones.cpp
@@ -1,16 +1,15 @@
 #include <benchmark/benchmark.h>
 
-#include <sim_sycl.h>
 #include <sim_seq.h>
+#include <sim_sycl.h>
 
 #include <random>
 
-using U = double;
+using U = float;
 
-static const simulation_configuration<U> config{.use_cutoff=false, .n_symetries=27};
+static const simulation_configuration<U> config{.use_cutoff = false, .n_symetries = 27};
 
-template<typename T, class ForwardIt>
-static inline void rand_fill_on_host(ForwardIt first, ForwardIt last, T min, T max) {
+template<typename T, class ForwardIt> static inline void rand_fill_on_host(ForwardIt first, ForwardIt last, T min, T max) {
     std::mt19937 engine(0);
     auto generator = [&]() {
         std::uniform_real_distribution<T> distribution(min, max);
@@ -19,23 +18,22 @@ static inline void rand_fill_on_host(ForwardIt first, ForwardIt last, T min, T m
     std::generate(first, last, generator);
 }
 
-template<typename T>
-static inline std::vector<coordinate<T>> generate_particules(size_t size, T spacing) {
+template<typename T> static inline std::vector<coordinate<T>> generate_particules(size_t size, T spacing) {
     std::vector<coordinate<T>> out(size);
     rand_fill_on_host<T>(out.begin(), out.end(), 0., std::pow(size, 0.33) * spacing);
     return out;
 }
 
 
-void lennard_jones_sequential(benchmark::State &state) {
+void lennard_jones_sequential(benchmark::State& state) {
     auto size = static_cast<size_t>(state.range(0));
     auto vec = generate_particules<U>(size, 10 * config.r_star_);
     size_t processed_items = 0;
     U last_energy = 0;
-    run_simulation_sequential(vec, config); // preheat ?
+    run_simulation_sequential(vec, config);   // preheat ?
     for (auto _: state) {
         auto start = std::chrono::high_resolution_clock::now();
-        auto[field, sums, energy] = run_simulation_sequential(vec, config);
+        auto [field, sums, energy] = run_simulation_sequential(vec, config);
         auto end = std::chrono::high_resolution_clock::now();
 
         processed_items += size * size * config.n_symetries;
@@ -47,7 +45,7 @@ void lennard_jones_sequential(benchmark::State &state) {
     state.SetItemsProcessed(static_cast<int64_t>(processed_items));
 }
 
-void lennard_jones_sycl(benchmark::State &state) {
+void lennard_jones_sycl(benchmark::State& state) {
     auto size = static_cast<size_t>(state.range(0));
     sycl::queue q{sycl::default_selector{}};
 
@@ -57,7 +55,7 @@ void lennard_jones_sycl(benchmark::State &state) {
     auto forces_device = std::span(sycl::malloc_device<coordinate<U>>(particules_host.size(), q), particules_host.size());
 
     size_t processed_items = 0;
-    auto[summed_forces, energy] =  run_simulation_sycl_device_memory(q, particules_device, forces_device, config);
+    auto [summed_forces, energy] = run_simulation_sycl_device_memory(q, particules_device, forces_device, config);
     for (auto _: state) {
         auto start = std::chrono::high_resolution_clock::now();
         run_simulation_sycl_device_memory(q, particules_device, forces_device, config);
